Feat: 진도 스테이트 추가

+++ code/FourQuest/FQClient/VibrationState.h
@@ -24,7 +24,7 @@ namespace fq::client
 		bool mbUseAllController; 
 		EVibrationMode mMode;
 		float mDuration;
-		float mIntensity;
+		int mIntensity;
 		float mElapsedTime;
 		float mVibrationTime;
 		unsigned int mControllerID;
